Fix: Split test case
