+++ test/Integration/Command/NormalizeCommand/Json/NotValid/fixture/composer.json
