+++ test/Integration/Command/NormalizeCommand/Json/Valid/Lock/NotPresent/Json/NotYetNormalized/fixture/composer.json
@@ -1,0 +1,13 @@
+{
+  "type": "library",
+  "_comment": "This composer.json is valid according to a lax validation, , a composer.lock is not present, and composer.json is not yet normalized.",
+  "keywords": ["foo", "bar"],
+  "license": "MIT",
+  "authors": [
+    {"name": "Andreas Möller", "email": "am@localheinz.com"}
+  ],
+  "require": {
+    "ext-json": "*",
+    "php": "^5.6"
+  }
+}
