+++ test/Integration/Command/NormalizeCommand/Json/Valid/Lock/Present/NotFreshBefore/WithoutNoCheckLock/fixture/composer.json
