+++ test/Integration/Command/NormalizeCommand/Normalizer/Throws/fixture/composer.json
